+++ src/app/components/Navbar/navbar.tsx
@@ -1,31 +1,21 @@
-import {
-  NavContainer,
-  StyledModeBtn,
-} from "@/app/styling/components/Navbar/styled.navbar";
-import { useAppDispatch, useAppSelector } from "@/app/redux/hooks";
-import { selectDarkMode, toggle } from "@/app/redux/features/darkModeSlice";
-import { DarkModeIcon, Logo } from "../SVGs/index";
+import { useAppSelector } from "@/app/redux/hooks";
+import { selectDarkMode } from "@/app/redux/features/darkModeSlice";
 import NavLinks from "./NavLinks";
 import Search from "./Search";
 import CurrencySelect from "./CurrencySelect";
+import DarkModeButton from "./DarkModeButton";
+import { Logo } from "../SVGs/index";
+import { NavContainer } from "@/app/styling/components/Navbar/styled.navbar";
 
 const Navbar = () => {
-  const dispatch = useAppDispatch();
   const { darkMode } = useAppSelector(selectDarkMode);
-
-  const toggleDarkMode = () => {
-    dispatch(toggle(darkMode));
-  };
-
   return (
     <NavContainer>
       <Logo darkMode={darkMode} />
       <NavLinks darkMode={darkMode} />
       <Search />
       <CurrencySelect darkMode={darkMode} />
-      <StyledModeBtn onClick={toggleDarkMode}>
-        <DarkModeIcon darkMode={darkMode} />
-      </StyledModeBtn>
+      <DarkModeButton darkMode={darkMode} />
     </NavContainer>
   );
 };
